Desafios/Projeto_Whisper_Python/requirements.txt
openai
python-dotenv
gtts
soundfile
sounddevice
numpy
scipy